refactor(cli): move studio CLI out of cli/, fold micro-helpers into _util
Second consolidation pass: studio-dedicated commands should not sit as
standalone files inside the general CLI package.

- `li studio` + `li schedule` move out of cli/ into lionagi/studio/cli.py:
  studio's CLI now lives with studio. cli/studio.py + cli/schedule.py deleted;
  main.py imports from lionagi.studio.cli. studio/__init__.py stays empty so
  base-install dispatch of `li studio`/`li schedule` remains FastAPI-free.
- micro-helpers cli/_lifecycle.py (27) + cli/_process.py (55) fold into one
  leaf cli/_util.py: both are shared seams (lifecycle by agent/state/orchestrate,
  process by monitor/kill + studio/services/admin), so a leaf util fits the
  dependency shape better than folding into a command file.

Behavior, HTTP route table, and `li` command surface unchanged. Rewires main.py
plus 8 source importers and 6 test files. Net -786 LOC from cli/, 4 files
removed, 2 added.

Gates: HTTP-contract oracle byte-identical (79 routes); full pytest green
(0 failed, 1 pre-existing skip); ruff check + format clean; opus critic APPROVE.
