+++ lionagi/cli/_util.py
@@ -1,11 +1,33 @@
 # Copyright (c) 2023-2026, HaiyangLi <quantocean.li at gmail dot com>
 # SPDX-License-Identifier: Apache-2.0
-"""Shared CLI process utilities: PID liveness and entity resolution."""
+"""Shared CLI utilities: exit-code mapping, exception classification, PID liveness, entity resolution."""
 
 from __future__ import annotations
 
 import os
 from typing import Any
+
+EXIT_CODE_BY_STATUS: dict[str, int] = {
+    "completed": 0,
+    "failed": 1,
+    "timed_out": 124,
+    "aborted": 130,
+    "cancelled": 143,
+}
+
+
+def classify_exception(exc: BaseException) -> str:
+    from lionagi._errors import TimeoutError as LionTimeoutError
+
+    if isinstance(exc, KeyboardInterrupt):
+        return "aborted"
+    if isinstance(exc, (TimeoutError, LionTimeoutError)):
+        return "timed_out"
+    from lionagi.ln.concurrency.errors import cancelled_exc_classes
+
+    if isinstance(exc, cancelled_exc_classes()):
+        return "cancelled"
+    return "failed"
 
 
 def pid_alive(pid: int) -> bool:
