+++ lionagi/cli/main.py
@@ -8,6 +8,13 @@ import argparse
 import signal
 import sys
 
+from lionagi.studio.cli import (
+    add_schedule_subparser,
+    add_studio_subparser,
+    run_schedule,
+    run_studio,
+)
+
 from ._logging import configure_cli_logging, log_error
 from .agent import add_agent_subparser, run_agent
 from .casts import add_casts_subparser, run_casts
@@ -20,10 +27,8 @@ from .orchestrate import (
     inject_playbook_schema_into_parser,
     run_orchestrate,
 )
-from .schedule import add_schedule_subparser, run_schedule
 from .skill import run_skill
 from .state import add_state_subparser, run_state
-from .studio import add_studio_subparser, run_studio
 from .team import add_team_subparser, run_team
