+++ lionagi/cli/orchestrate/flow.py
@@ -15,10 +15,10 @@ from lionagi.casts.emission import SpawnRequest
 from lionagi.ln.concurrency import CancelScope, move_on_after
 from lionagi.orchestration import plan, role_node_builder
 
-from .._lifecycle import classify_exception
 from .._logging import progress
 from .._logging import warn as _warn
 from .._providers import parse_model_spec
+from .._util import classify_exception
 from ._common import (
     _create_fanout_team,
     _format_result_json,
