+++ tests/apps_studio_server/test_startup_warnings.py
@@ -403,7 +403,7 @@ class TestCLIHostWiring:
     """
 
     def test_backend_only_exports_resolved_host(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        import lionagi.cli.studio as studio_cli
+        import lionagi.studio.cli as studio_cli
 
         monkeypatch.delenv("LIONAGI_STUDIO_HOST", raising=False)
         monkeypatch.setattr(studio_cli, "_ensure_apps_importable", lambda: True)
@@ -432,7 +432,7 @@ class TestCLIHostWiring:
         self, monkeypatch: pytest.MonkeyPatch, tmp_path: Path
     ) -> None:
         """_start_local must overwrite a stale LIONAGI_STUDIO_HOST before uvicorn.run."""
-        import lionagi.cli.studio as studio_cli
+        import lionagi.studio.cli as studio_cli
 
         # Stale env claims 0.0.0.0; the CLI is about to bind 127.0.0.1.
         monkeypatch.setenv("LIONAGI_STUDIO_HOST", "0.0.0.0")
